Add files via upload
Added missing fields to RecipeFrame.  All class member variables are now update-able in the application via member functions.
RecipeList was loading an incorrect value for one of the fields, it now loads the correct value.

+++ Units.cpp
@@ -26,7 +26,7 @@ bool doesTypeMatchForConversion(Type from, Type to)
 	return false;
 }
 
-float standardUnitConversion(Unit start, Unit end, float value)
+double standardUnitConversion(Unit start, Unit end, double value)
 {
 	// increment if we are going up in scale, ounce to pounds
 	while (start != end)
